Decoupling crypto digest to choose crypto algos
In this commit we will create new classes such that the Digest data and
digest creation is decoupled. Digest data will be kept at the same
place but the creation can be choosen by different libraries like
cryptopp or openssl. This rearrangement of classes will be done in this
PR.